Write output directly to disk with -o
This prevents caching the entire output string in memory, reducing the heap usage for our largest output file of 7.2 MB from 101 MB to 93 MB in my test.

Only JSON is supported for now. YAML requires more work because the YamlRenderer is more closely tied to the underlying StringWriter, so it still has to buffer.