Fix import caching
This introduces `Importer` as another layer below `Resolver` to separate the translation of a relative import path into an absolute path from reading the file. So far we always read the same file again if it was imported multiple times during a single evaluation. We can improve both correctness and performance by caching this information. Note that the global parse cache which is shared across multiple evaluations when using our Bazel JsonnetWorker does not contain this new data. It is only kept locally in the `CachedResolver` because it is not valid for multiple evaluations.