More conventional scope handling
This is one of the most consequential changes to the design of the interpreter so far. Previously all variables in the `ValScope` objects that are used during evaluation were stored with a name-based index, using different indices per source file. The new scheme is symbol-based, i.e. when a name gets shadowed we do not override the location of the previous definition in the `ValScope`, instead we always append new definitions at the end. This has several advantages:
- All scopes start empty at the root of an AST, they only grow as more definitions are bound when going deeper. The old scheme required that we allocate the maximum length for the bindings array at every point because definitions could end up at any index.
- The handling of bindings in function calls becomes a lot simpler. It is no longer necessary to shuffle parameters with a translation table from parameter position to name-based index. They are simply appended en bloc at the end of the bindings.
- We can efficiently detect when a default value is needed and when a parameter is bound multiple times. The old `BitSet`-based validation is no longer necessary.
- Name-based parameters can be used in function calls across source files without having to translate the indices.

The old name index assignment is removed from the parser. We simply keep the original strings in the AST. This also removes the need for the mapping table in the `FileScope`. String-based names are always available for use in error messages without any lookups.

Since Jsonnet allows forward references, the parser cannot assign symbols or symbol-based indices. Instead this is now handled by the `StaticOptimizer` which translates `Id` expressions into `ValidId` containing the index obtained from the static scope. Definitions (like `Bind` or `ForSpec`) do not contain indices because they always append new bindings at the end. All references to the standard library (which used to be in slot 0 of the scope in the old name-based design) are replaced by the static `Std` object itself, so no lookup is necessary during evaluation.

`self`, `super` and `$` still use the name-based lookup with their special slots in the `ValScope` for now. This will be changed later.