Val extends Lazy
`Val` can finally extend `Lazy`. I tried this before in several variations but the effect on performance was always negative. With the other changes so far we can finally make it work as intended. This allows us to remove `Val.Strict` and treat every `Val` directly as a `Lazy`. It requires an additional reference in each `Val` for the cached value (which is essential to benefit from the monomorphic `force` call) but in return we can avoid allocating many unnecessary `Strict` wrappers.