Optimize Materializer + Renderer
These optimizations are based on ujson 1.3.7. We previously used a vendored version of `BaseRenderer` as the basis for our renderers. No we are using the newer `BaseCharRenderer` which offers much better performance, reducing the time for the materializer benchmark from >60ms to 15ms. The very complex `PrettyYamlRenderer` has not been updated yet.