More Apply optimization
This extends the previous optimization to `Apply0` and adds a new optimization for rebinding calls to functions that can be statically resolved. If we know the target of a function call we can resolve named arguments and add default values as necessary to turn any function call into a "simple" one that does not require named arguments and defaults to be handled during evaluation.