Remove special handling of self/$/super in the symbol table
We now treat `self` and `super` the same way as function parameters. An object definition reserves two indices in the `ValScope` and evaluating the object supplies the values. `$` for the outermost object does not require a place of its own. Since we are not shadowing `self` anymore, a reference to `$` can still refer to the outermost `self` at its original index. All uses of `Expr.Self` and `Expr.$` are replaced by `ValidId` in the optimizer. `Super` requires a special `ValidSuper` type.