Builtin with defaults & std.setInter optimization
Built-in functions with defaults were still treated as `Val.Func`. With the old scope handling we would have needed a more complex implementation to also handle default values but this is no longer a problem. We simply have to pass them on to the superclass.

Now that we can turn `std.setInter` into a `Builtin` we can optimize it further by partially applying a static argument.