Add support for warnings and --no-static-errors
Warnings could prove useful in the long run if we want to add linting. `--no-static-errors` is provided for migrating from the previous version but should be removed again in the long term because the specification requires static error handling. Since warnings are not thrown they do not get a full stack trace. Only the actual position of the warning is reported.