Add --fatal-warnings
Now that we have warnings, we need an option to make them fatal. Note that the combination of --no-static-errors --fatal-warnings is not the same as running without any of these flags. Static errors are detected after parsing a source file. Since this is a lazy process, any additional sources that would have been needed for evaluating the file are not loaded and thus not checked for errors. Running with --no-static-errors --fatal-warnings ensures that all static errors are reported and the interpretation fails (when the code paths leading to the errors would not have been evaluated otherwise).