Fix performance regression in split by avoiding allocating substring per char (#237)
This PR fixes a performance regression from #227 /
4c85bde425b51103a95efce272fa0e4fb1bd51dd which I overlooked in review:

When generalizing the optimized non-Pattern-based split code, that
commit introduced a `.substring()` on each character, producing tons of
garbage.

Instead, I think we can do a `.startsWith(splitPattern, i)`: this should
be much faster because it will avoid unnecessary garbage string creation
(plus I'm pretty sure that `startsWith` is optimized in modern JDKs).

I also removed the use of `breakable` and replaced it with an update to
the `while` condition.