Add support for tailstrict (#257)
With this PR, I'm implementing tailstrict in sjsonnet.
I'm following the guidance from
https://github.com/google/jsonnet/issues/343#issuecomment-325987638:

Quote:
1. If you call a function with tailstrict annotation on the apply AST,
e.g. foo(42) tailstrict then the evaluation of the function body happens
in "tail strict mode". The annotated AST need not actually be a tail
call. Also, the arguments of this function are forced.
2. When another function call (can be a completely different function to
the one that was originally called) is made when we're evaluating in
"tail strict mode" and it is a tail call, then the current stack frame
is re-used for the next call, rather than being pushed on top of.
3. Note that in order to preserve the tail strict mode into the new
function, the new call AST has to be tailstrict as well.

Resolves #189.