Allow `std.map` to operate on string args as well.
In gojsonnet, a few functions of the stdlib accepts strings in addition
to arrays. Sometimes this is mentioned in the stdlib documentation, sometimes
it isn't - and importantly, this is a special property of a few select
functions and it does NOT mean that a string can be treated as an array
of single-character strings in every context with implicit type
conversion between the two.

Eyeballing the gojsonnet codeyields the following list of std functions
which should support strings as parameters:

std.slice
std.member
std.map
std.flatMap
std.base64

Out of those, only `map` had missing support for strings, which is being
added in this commit.

In the process, I discovered other quirks of the `map` function, such as
the undocumented, `filterMap`-like behavior from this test case:

`std.assertEqual(std.map(function(x) x * x, std.filter(function(x) x > 5, std.range(1, 10))), [36, 49, 64, 81, 100])`