fix(js,native): std.parseYaml() handling of "---".
Problem: On js and native targets std.parseYaml() mishandles input where
"---" is not at the beginning of the line. (The jvm target is fine.)

Fix: Only match "---" at the beginning of the line. Also don't require a
newline after it.

Note that there are other corner cases that are still broken, but for
those I'll file a bug against the underlying scala-yaml library.